+++ CHANGELOG.md
@@ -19,8 +19,10 @@
   - Now it uses CIM cmdlets to get information from WMI classes.
   - Resolved all of the PSScriptAnalyzer warnings that was triggered in the common tests.
 - Changes to xSQLServerScript
-  - All credential parameters now also has the type [System.Management.Automation.Credential()] to work with PowerShell 4.0.
+  - All credential parameters now also has the type [System.Management.Automation.Credential()] to better work with PowerShell 4.0.
   - It is now possible to configure two instances on the same node, with the same script.
+  - Added to the description text for the parameter `Credential` describing how to authenticate using Windows Authentication.
+  - Added examples to show how to authenticate using either SQL or Windows authentication.
 
 ## 4.0.0.0